requirements.txt
requests
auth
pyjwt
dataclasses
pathlib
pyyaml